robot/requirements.txt
pyfirmata
numpy
opencv-python
fastsam
torch
torchvision
Pillow
requests
scipy
requests_oauthlib
tqdm
